split indy user to wallet service and ledger service, refactor roles
